Refactor ClientRecord to store AssignedHunts over multiple records
This ensures the document does not grow too much. Previously
AssignedHunts were stored in a single array in the same document and
we overwrote the doc id.

Now documents are merged dynamically to figure out which hunts were
assigned. Additionally the foreman now only queries for active hunts
reducing the amount of work required.