ENDPOINT-7506: Fix for duplicate clients appearing in search (#127)
## Description
When you filter for online clients and then search ( for a hostname for
example ), duplicate clients are shown. I updated the searchVerbs
function so that it will only check for a label with the search term
after zero hostnames with that search term are returned as this was
causing the duplication.

Co-authored-by: Aoibhe Gray <aoibhe_gray7@rapid7.com>