Fixed bugs in document ids
The issue was that before this PR, we had multiple indexes which
stored different types of documents and for some indexes the same
document id was used (e.g. acls, users and user options were all using
the username as the doc id). However this PR changed all the index
names to "persistant" which means that now all those different records
were overwriting the same doc ids in the same index.

Simple testing would have found these issues so we need to make sure
to test a bit better.

I am not 100% sure that this is fully fixed - I suggest we review all
the transformations and ensure that all SetElasticIndex() with
non-random id are properly deduplicated to avoid these overwrites.