Bugfix: LoadCollectionContext only merged some items.
Since migration to the datastream indexes, collection context is
updated many times through the collection lifetime. To reconstruct the
most up to date object we need to scan all the rows (since they are no
longer overwritten).

This PR fixes a bug where only a few collection documents were fetches
so for collections which dont return results for a while (but still
send progress messages) the GUI was unable to see the latest state of
the object and did not show any results.