Refactor result sets.
Since moving to transient index it is no longer possible to delete
result sets. So now when we need to overwrite them we have to create a
new result set each time.

This refactors associates an ID with each result set and rewrites it
from scratch when the writer truncates it.