Implemented lazy hunt flow index updates (#166)
Previously this was done on each access to hunt_dispatcher.GetFlows()
which presented a lot of overheads. We now compare the last modified
time of the result set and the hunt updates to see if the hunt flow
index needs to be rebuilt.

Also fixed implementation of hunt_add() since cloudvelo does not have a
hunt_manager, it was not able to handle any of the mutations sent by the
default VQL function.